Create basic Rails app template
The app was generated from a template created with the rails_apps_composer gem. Some preferences were predefined in a defaults YAML file, and the Gemfile was created via a custom recipe.

My goal is to refine this template based on best practices so that future Rails apps can be created from it.