Add basic authentication with Devise
This helped a lot: http://railsapps.github.io/tutorial-rails-mongoid-devise.html

Things that weren't covered in that tutorial: how to set up smtp for Gmail and how to set up the test environment for testing the mailer. See the changes in config/environments/development.rb and config/environments/test.rb

Also, I had to rework the inheritance in the controllers because Devise won't work if application_controller.rb is inheriting from RocketPants::Base.