Make 'api_applications/:id' go straight to edit action
The Rails scaffold generator creates a few unnecessary actions and files. One of them is the "show" action that just shows the information. A user must then click the "Edit" link to be able to update the application. I think it's a better experience for the user to be taken directly to the edit mode when they click on one of their applications.