Create initial admin interface for all resources with ActiveAdmin
ActiveAdmin does not support Mongoid, so we're using activeadmin-mongoid, which still has many issues. I got most of what we need working. One thing that doesn't work, but which I don't foresee needing, is the ability to edit a user's api application.

To use filters in the admin interface, I had to use a hack that someone else wrote, which I replicated in my fork of activeadmin-mongoid.