Add pagination headers
I think this is easier for clients to use than defining new methods in the wrapper. The best place for pagination methods in the wrapper would be as a method of @last_response, but that would mean adding the methods to Sawyer.

Clients that are using Ohanakapa can get pagination info by doing this:
locs = Ohanakapa.locations
response = Ohanakapa.last_response
current_page = response.headers["X-Current-Page"]