Only geocode if address has changed
When I turned the after_validation callback back on, I didn't realize that the geocoding service would be called every time a location was updated, even if it already had coordinates. This was unnecessarily affecting the limited amount of requests that Google give us.

Now, it only geocodes if the physical or mailing address have changed.