Only the admin app should be allowed to write to the API
Since the API is serving live data, we can't allow everyone to make changes to the data. Only the admin app that we built should have access. If down the line someone wants to build a better admin interface, then they can get approval from the city/county/community that owns the data.