Remove farmers' markets from DB setup task
This is necessary for the Ohana API Admin specs to pass.

Farmers' Markets were specific to SMC. It was just an additional data source we decided to add, but it's not part of the OpenReferral schema.

There are still some references to markets, namely within Kind, which we will deal with later. The question is whether or not the Kind field would be useful to other cities since the assumption is that all OpenReferral data is of the Human Services kind.