Use db:structure:load instead of db:schema:load
Because we have migrations that ActiveRecord can't handle, we had to change our schema format to `sql` in application.rb.
This means we should now use db::structure:load to create the DB, so that it uses structure.sql. Otherwise, if we kept using db:schema:load, it would complain that there are pending migrations — the ones that execute SQL that ActiveRecord can't perform.