Replace hyphens with underscores in DB names
The PG DB naming convention is to use lowercase and underscores, but you can use hyphens, as long as you surround the DB name in quotes. Rails automatically adds the quotes for you, but Travis does not. So, to make it easier and consistent all around, I just used underscores everywhere.