Add wait time after deleting orgs/locs/services
Intermittently, the specs that delete a location, organization, or
service fail because the driver is checking the URL before it has had a
chance to update.

This commit adds a 1-second wait time to allow the page to update. The
wait time can be increased if necessary.

I also added a spec for deleting services since it was missing.