Only require service name & desc for human service
Since SMC data has other kinds of organizations where the concept of "service" doesn't apply, the service name and description should not be required.

For those non-human services locations, it might also be better to display the "service" fields in the edit location form instead of having to click a link to edit the service since there is only one "service" in most cases.