Set Cache-Control to max-age=300, public
To improve the performance of Ohana Web Search, and to reduce the load on the API, we tell Ohana Web Search to cache API responses for a specified amount of time.

The 2 API calls that Ohana Web Search makes correspond to the `index` action in `search_controller` and the `show` action in `locations_controller`. By default, Rails sets Cache-Control to `max-age=0, private, must-revalidate`. By adding the `expires_in` line to the end of the controller actions, we set Cache-Control to `public`, and set `max-age` to the number of minutes specified in the `EXPIRES_IN` env var.

In Ohana Web Search, this is handled by the Ohanakapa Ruby wrapper in conjuction with the `faraday-http-cache` gem, which determines whether or not to repeat an API request based on the Cache-Control header in the API response.