Specify subdomain for Devise mailers.
Since we have both a developer and admin portal, we need to make sure
that the links provided in the emails sent by Devise (such as when a
user signs up or asks to reset their password) point to the correct
portal.

I also cleaned up the localization and copy for the portals.