Use Select2 for service areas field.
Since the list of valid service areas is defined in settings.yml, it
should be used to populate the choices in the select field in the admin
interface. That way, it won’t be possible to submit an invalid value,
and users won’t have to guess what to enter.