Disable unicorn worker killer.
There seems to be a memory leak in the app, and unicorn worker killer
is supposed to help manage that by restarting unicorn workers that
exceed a certain memory threshold. That part is working, but when a
worker is restarted, it’s causing H13 errors on Heroku.

If memory increases past the dyno limit, Heroku will raise R14 errors,
but the app still continues to serve responses, only slower. I’d prefer
to get R14 errors instead of H13, so I’m disabling unicorn worker
killer for now.