Add Organization fields per v0.2 of HSDS
This commit adds the following fields to the Organizations table:

- alternate_name
- date_incorporated
- description
- email
- legal_status
- tax_id
- tax_status
- website

The `urls` field has been removed because it has been replaced by the
singular `website` field.

All of the new fields are exposed by the API except `legal_status`,
`tax_status`, and `tax_id`.

All of the new fields can be updated or added to a new organization via
the admin interface or via writing to the API.

The description field is required per the OpenReferral spec.

The following fields defined in the spec have not been added because
it’s not clear whether or not they belong in the Organizations table or
the Locations table:

- license
- accreditation
- funding