Remove unused custom pagination HTTP headers.
Only `Link` and `X-Total-Count` are useful for clients such as Ohana
Web Search. Originally, the others (such as `X-Total-Pages` and
`X-Current-Page`) were added to help implement pagination in Ohana Web
Search, but we realized that we could just use existing tools such as
the `kaminari` gem to do the pagination for us.