Drop Faxes table & add fields to Phones
Instead of having a separate table for Faxes, we can use the
`number_type` field in the Phones table to identify a fax.

The `number_type` field, which is now required, accepts one of four
values:
- fax
- hotline
- tty
- voice

The optional `country_prefix` field has been added to the Phones table.
Valid values are country dialing codes, including the `+`, such as
`+212` for Morocco.