Optimize nearby endpoint.
The most common way the nearby endpoint would be used is to display
locations near the one currently being visited on a map. A client would
only need a few location attributes to make use of this endpoint, such
as the name, alternate name, latitude, longitude, address, and slug.

By creating a special serializer that only returns a few attributes, we
improve the performance of the nearby endpoint.