Expose regular_schedules in GET location endpoint.
`opens_at` and `closes_at` are presented unmodified in ISO 8601 format.
`weekday` is stored as an integer in the database, but is presented as
the English weekday name, such as “Monday”.

This commit also changes the integer representing Sunday from 0 to 7 so
that schedules can be ordered starting with Monday.