Update DB setup scripts and documentation.
Use the PG dump file to restore the DB during `script/bootstrap`
instead of running a bunch of scripts.

Create new scripts for exporting and importing the DB for production
use, and gitignore this production dump for security.

Update CONTRIBUTING.md with instructions for updating sample data and
pushing to Heroku apps.

Closes #258.