Add holiday_schedules table.
This table is used to identify the dates when a location or service is
either closed, or open with modified hours.

Required fields: `closed`, `start_date`, `end_date`.

Valid values for `closed` are: 0, ‘0’, false, ‘false’ to denote that
it’s open; and 1, ‘1’, true, ‘true’ to denote that it’s closed.

If `closed` is false, then `opens_at` and `closes_at` become required
fields.

`start_date` and `end_date` are fields of type Date, but can accept a
date in String format, as long as it contains a year, and is formatted
like this: “December 24, 2014”.