Limit status search results to one page.
If pagination is not added to the query, it will load all results in
memory. Since the status controller is just meant to ping the API to
make sure it’s alive, the search should return as few results as
possible.

Adjust the actual keyword searched here to match what’s in your
database.