Add date validator.
This allows dates to be submitted in a few different formats, namely
month/day/year, or spelled out, as in November 17, 2014.

In the month/day/year format, both 2-digit and 4-digit years are
accepted.

The day/month/year format is also accepted, but to use it, you must
first update the `date_format` option in `settings.yml`.