Update DB in preparation for merging upstream code
Run the necessary migrations to update the DB to match the Open
Referral spec, but without removing or renaming old fields because
SMC-Connect depends on those old fields.

The content from the old fields has been copied over to the new fields,
and both old and new fields will be exposed by the API to allow
SMC-Connect to update.

Once SMC-Connect is updated and using the new fields, we can remove the
old fields from the API.