Ignore CSV files that are not required.
If a CSV file that is not required is either missing or present  but
empty, the import script will skip it and won’t raise an error.

If a required file is missing or empty, then the rake task will be
aborted and an error message will appear explaining that the file is
missing or empty.

Closes #284.