Add support for importing taxonomy via CSV.
This allows categories to be created in a tree structure supported by the ancestry gem. To see an example of a valid CSV file that recreates the Open Eligibility taxonomy, look at data/taxonomy.csv

For more details about how taxonomy works, read our Wiki article on taxonomy basics:
https://github.com/codeforamerica/ohana-api/wiki/Taxonomy-basics

Special thanks to @volkanunsal for the feature request and getting this work started.