Enhance admin interface dashboard UX
If there aren't any organizations in the database yet, the "Add a new location" and "Add a new program" buttons will be hidden because locations and programs require a parent organization.

If no organizations exist, and a super admin visits the "create a new location" form directly by entering the URL, they will be granted access instead of getting a "not authorized" error.