Fix bug with language search.
If the search also contained the category parameter, Postgres would complain about the languages column being ambiguous. That's because both the Services and Locations table have a languages column now, and since the category parameter includes a join with the Services table, we need to specify that the languages column is in the Locations table.

Eventually, we should probably include the Service languages column in the search as well.