Move sample CSV files to a separate folder.
The sample CSV files are there only as an example. None of them are
actually used by the demo app.

Leaving them in the root data folder would cause sample data to be
imported when running the import script, which might not be desirable
when you have your own data.

Closes #296.