Display Ohana API version for super admins.
The version number is based on the VERSION file from the codeforamerica/ohana-api repo. Because deployments can manage and customize their instances of Ohana API however they choose, the version number does not necessarily mean that a particular deployment is exactly in sync with the upstream repo. It should be used as a general indicator of when an update was last peformed.

The version number appears in the footer of the admin home page, and only for super admins. It also links to the ohana-api repo's Changelog to make it easy to see what has changed since a particular version.

Closes #294.