Provide one-click download of CSV files as a zip
Since large files can take a while to process, the zip file is
prepared in the background so that the user can continue browsing
the site. The background processing is provided via the Sucker Punch
gem, which doesn't require any additional Heroku dependencies such as
Redis.

While the zip file is being generated, the button on the admin
interface reads "Generating zip file", and when it's ready, it reads
"Download zip file".

Once the file has been downloaded, it is deleted from the /tmp folder
a minute later so that a new zip file can be generated at a later time.

The zip file's name contains the date at which it was downloaded to
make it easy to see when the zip was last generated.

I also cleaned up the csv.shaper files to remove duplicated cells.