Disable email delivery during rake db:seed
Because we are using the letter_opener gem, every time an email gets sent, a new browser tab is opened. While testing and resetting the DB, having browser tabs open each time and steal focus from the Terminal gets annoying.

We're not testing or using emails during the seeding of the DB, so we can suppress emails by using Devise's `skip_confirmation!` method.