Update sample csv files
Now that we have an export feature, I exported the demo data as CSV files and used that to replace the old sample CSV files, which were manually created with dummy data.

I also fixed the accessibility CSV export so that it uses the format that's expected when importing the data.