Replace pg_search with activerecord query
Why:

We're only using the pg_search gem for the keyword search, and the
query that it generates is not that complicated. Upgrading to the
latest version of the pg_search gem breaks our tests. So, instead of
having to rely on a third party, we can generate the same query
ourselves using activerecord, and reduce our dependencies with the
same stone.

How:

To figure out what the proper query was, I fired up the Rails console
and ran `Location.keyword('food')`. That gave me the SQL query I
needed, which I then converted to activerecord so that the result
would return an ActiveRecord::Relation that can be chained with other
scopes.