Order results by updated_at when rank is the same
Why:
Search queries that contain the keyword parameter are ordered by the
rank calculated by the Postgres full text search feature. In some
cases, the rank could be the same for all results, but in those cases,
Postgres does not guarantee that the order will be same all the time.
This will cause issues with pagination since the order might not be
consistent as you go from one page to another.

In order to keep the order consistent, we can specify an additional way
to rank the results. I've chosen to order by most recently updated.