Eliminate duplicates when using service_area
When performing a search, if a location has many services that match the `service_area` parameter, the same location will be duplicated in the search results.

To eliminate the duplication, you can either append `.uniq` or `.group('locations.id')` to the `Location#service_area` query. However, both are not without consequences. If the keyword parameter is also present when using `.uniq`, then we need to add a select clause that selects all location columns, as well as the full-text search rank.

If we use `group`, then we get an error when the `location` parameter is present due to the way the geocoder gem works. The issue in this case is caused by the `maximum(:updated_at)` call used in the `cache_key` method in `search_controller`. I couldn't figure out how to make `group` work, so I went with the former.