Disable rack-cache
Why:
It is known that using a store like Memcache or Redis for the rack-cache Entitystore uses up a lot of memory. Normally, it's recommended to use disk storage for the Entitystore, but that is not possible on Heroku. For some reason, this current implementation is using up more memory than previously (at least that's what it looks like).

For now, disabling rack-cache is the way to keep memory usage at norma levels.