Remove new_relic requires & optimize controllers
When I previously removed the New Relic method tracing, I forgot to remove the require statements, and I think that's contributing to the memory leak.

While I was in there, I added some optimizations to use memoization for constant values like the DB tables used for preloading and includes.

I also updated the geocoder config to use the latest syntax in the readthis gem.