Add CSV download links for all tables
**Why**: After I made the app more secure by only allowing super admins
to access CSV downloads, I inadvertently broke the zip file generation.

I'm not sure there's a way to restrict access while allowing a
background job to dynamically generate and fetch the CSVs. Instead, I
opted to revert back to having individual download links for all
tables, which is what this app originally had a while back.

Looking at the PR that replaced the individual links with the zip file,
it looks like one of the reasons was because large files can take a
while to download. To keep things simple in this PR, I simply added a
note on the admin dashboard that lets people know large files can take
a while.

Fixes #437