Update Rails from 4.2.10 to 5.0.6
This makes the minimum amount of changes to allow the tests to pass.
There are deprecation warnings that will be addressed in a subsequent
PR.

The biggest change in this PR is converting the `admin_emails` field
from a serialized text field to a native Postgres array column.
This was necessary because the old Arel queries in EmailFilter no
longer work, and switching to an array column was the easiest way to
fix the issue. This change is a good idea indpendent of the Rails
upgrade for performance reasons in addition to the easier to read
queries.